// app/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      inventory: {
        Row: {
          back_orders: number | null
          created_at: string
          inventory_id: string
          inventory_level: Json | null
          product_id: string | null
          product_image: string | null
          product_name: string | null
          sku: string | null
          store_url: string
          variant_id: number | null
          variant_name: string | null
        }
        Insert: {
          back_orders?: number | null
          created_at?: string
          inventory_id: string
          inventory_level?: Json | null
          product_id?: string | null
          product_image?: string | null
          product_name?: string | null
          sku?: string | null
          store_url: string
          variant_id?: number | null
          variant_name?: string | null
        }
        Update: {
          back_orders?: number | null
          created_at?: string
          inventory_id?: string
          inventory_level?: Json | null
          product_id?: string | null
          product_image?: string | null
          product_name?: string | null
          sku?: string | null
          store_url?: string
          variant_id?: number | null
          variant_name?: string | null
        }
        Relationships: []
      }
// ...
      order: {
        Row: {
          created_at: string
          currency: string | null
          current_total_tax: string | null
          current_total_tax_set: Json | null
          customer_email: string | null
          customer_id: number | null
          financial_status: string | null
          fulfillment_status: string | null
          fulfillments: Json | null
          id: string
          line_items: Json[] | null
          order_id: number
          order_number: number
          order_status_url: string | null
          order_tags: string | null
          processed_at: string | null
          shipping_address: Json | null
          shipping_costs: number | null
          shipping_costs_usd: string | null
          store_url: string | null
          sub_total_price: string
          sub_total_price_usd: string | null
          tax_rate: number | null
          total_discounts: string | null
          total_discounts_set: Json | null
          total_order_value: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          currency?: string | null
          current_total_tax?: string | null
          current_total_tax_set?: Json | null
          customer_email?: string | null
          customer_id?: number | null
          financial_status?: string | null
          fulfillment_status?: string | null
          fulfillments?: Json | null
          id?: string
          line_items?: Json[] | null
          order_id: number
          order_number: number
          order_status_url?: string | null
          order_tags?: string | null
          processed_at?: string | null
          shipping_address?: Json | null
          shipping_costs?: number | null
          shipping_costs_usd?: string | null
          store_url?: string | null
          sub_total_price: string
          sub_total_price_usd?: string | null
          tax_rate?: number | null
          total_discounts?: string | null
          total_discounts_set?: Json | null
          total_order_value: string
          updated_at: string
        }
        Update: {
          created_at?: string
          currency?: string | null
          current_total_tax?: string | null
          current_total_tax_set?: Json | null
          customer_email?: string | null
          customer_id?: number | null
          financial_status?: string | null
          fulfillment_status?: string | null
          fulfillments?: Json | null
          id?: string
          line_items?: Json[] | null
          order_id?: number
          order_number?: number
          order_status_url?: string | null
          order_tags?: string | null
          processed_at?: string | null
          shipping_address?: Json | null
          shipping_costs?: number | null
          shipping_costs_usd?: string | null
          store_url?: string | null
          sub_total_price?: string
          sub_total_price_usd?: string | null
          tax_rate?: number | null
          total_discounts?: string | null
          total_discounts_set?: Json | null
          total_order_value?: string
          updated_at?: string
        }
        Relationships: []
      }
      order_items: {
        Row: {
          created_at: string
          currency: string | null
          discount_allocations: Json | null
          grams: number | null
          id: string
          image_url: string | null
          lineItem_id: number | null
          order_id: number
          price: string | null
          product_id: number | null
          quantity: number | null
          sku: string | null
          store_url: string | null
          total_discount: string | null
          vendor_name: string | null
        }
        Insert: {
          created_at?: string
          currency?: string | null
          discount_allocations?: Json | null
          grams?: number | null
          id?: string
          image_url?: string | null
          lineItem_id?: number | null
          order_id: number
          price?: string | null
          product_id?: number | null
          quantity?: number | null
          sku?: string | null
          store_url?: string | null
          total_discount?: string | null
          vendor_name?: string | null
        }
        Update: {
          created_at?: string
          currency?: string | null
          discount_allocations?: Json | null
          grams?: number | null
          id?: string
          image_url?: string | null
          lineItem_id?: number | null
          order_id?: number
          price?: string | null
          product_id?: number | null
          quantity?: number | null
          sku?: string | null
          store_url?: string | null
          total_discount?: string | null
          vendor_name?: string | null
        }
        Relationships: []
      }
// ...
      stores: {
        Row: {
          created_at: string
          id: string
          is_inventory_fetched: boolean
          is_store_listed: boolean
          store_name: string
          store_url: string
        }
        Insert: {
          created_at?: string
          id?: string
          is_inventory_fetched: boolean
          is_store_listed?: boolean
          store_name: string
          store_url: string
        }
        Update: {
          created_at?: string
          id?: string
          is_inventory_fetched?: boolean
          is_store_listed?: boolean
          store_name?: string
          store_url?: string
        }
        Relationships: []
      }
      trackings: {
        Row: {
          created_at: string
          destination: Json | null
          id: string
          order_id: number
          shipment_status: string | null
          status: string
          store_location: string | null
          store_url: string | null
          tracking_company: string | null
          tracking_number: string | null
          tracking_numbers: Json | null
          tracking_url: string | null
          tracking_urls: Json | null
          updated_at: string
        }
        Insert: {
          created_at?: string
          destination?: Json | null
          id?: string
          order_id: number
          shipment_status?: string | null
          status: string
          store_location?: string | null
          store_url?: string | null
          tracking_company?: string | null
          tracking_number?: string | null
          tracking_numbers?: Json | null
          tracking_url?: string | null
          tracking_urls?: Json | null
          updated_at: string
        }
        Update: {
          created_at?: string
          destination?: Json | null
          id?: string
          order_id?: number
          shipment_status?: string | null
          status?: string
          store_location?: string | null
          store_url?: string | null
          tracking_company?: string | null
          tracking_number?: string | null
          tracking_numbers?: Json | null
          tracking_url?: string | null
          tracking_urls?: Json | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "trackings_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: true
            referencedRelation: "order"
            referencedColumns: ["order_id"]
          },
        ]
      }
// ...
    }
// ...
  }
}